+++ src/Plausible.php
@@ -11,6 +11,7 @@ namespace shornuk\plausible;
 use shornuk\plausible\services\PlausibleService;
 use shornuk\plausible\variables\PlausibleVariable;
 use shornuk\plausible\models\Settings;
+use shornuk\plausible\widgets\CurrentVisitors;
 use shornuk\plausible\widgets\TopPages;
 use shornuk\plausible\widgets\TopSources;
 use shornuk\plausible\widgets\TopDevices;
@@ -86,6 +87,7 @@ class Plausible extends Plugin
             Dashboard::class,
             Dashboard::EVENT_REGISTER_WIDGET_TYPES,
             function (RegisterComponentTypesEvent $event) {
+                $event->types[] = CurrentVisitors::class;
                 $event->types[] = TopPages::class;
                 $event->types[] = Overview::class;
                 $event->types[] = TopSources::class;