+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.5 - Unreleased
+
+#### Added
+- Current Visitors widget
+
+#### Fixed
+- Trailing and leading slashes in URI's. (https://github.com/shornuk/craft-plausible/issues/7)
+
 ## 1.0.4 - 2021-03-24
 
 #### Added